Rename snippet files for easier usage
