+++ widgets/src/lib/inferenceSnippets/inputs.ts
